+++ src/components/track-actions.ts
@@ -18,7 +18,7 @@ import {
 
 // our imports.
 import {
-  ALERT_ERROR_SPOTIFY_PREMIUM_OR_ELEVATED_REQUIRED,
+  ALERT_ERROR_SPOTIFY_PREMIUM_REQUIRED,
   ALERT_INFO_PRESET_COPIED_TO_CLIPBOARD,
   ALERT_INFO_PRESET_JSON_COPIED_TO_CLIPBOARD,
   RADIO_SEARCH_KEY
@@ -662,7 +662,7 @@ class TrackActions extends FavActionsBase {
 
         // spotify premium account (or elevated credentials) required for this function.
         if (!this.player.isUserProductPremium() && (!this.player.attributes.sp_user_has_web_player_credentials)) {
-          throw new Error(ALERT_ERROR_SPOTIFY_PREMIUM_OR_ELEVATED_REQUIRED);
+          throw new Error(ALERT_ERROR_SPOTIFY_PREMIUM_REQUIRED);
         }
 
         // have to hide the progress indicator manually since it does not call updateActions.